+++ app/Http/Controllers/Promotion1Controller.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Database\QueryException;
 use App\Models\HpProduct;
 use App\Models\HpPromotion;
 use App\Models\EcpayResult;
@@ -48,12 +49,19 @@ class Promotion1Controller extends Controller
 
     public function edit(HpPromotion $promotion1)
     {
-        $products = HpProduct::where('proj_id', 1)->where('status', true)->get();
-        $results = EcpayResult::where('trade_no', $promotion1->trade_no)->first();
-        $gifts = array();
-        if ($promotion1->gifts != null) {
-            $gifts = json_decode($promotion1->gifts);
+        try {
+              $products = HpProduct::where('proj_id', 1)->where('status', true)->get();
+              $results = EcpayResult::where('trade_no', $promotion1->trade_no)->first();
+              $gifts = array();
+              if ($promotion1->gifts != null) {
+                  $gifts = json_decode($promotion1->gifts);
+              }
+        } catch (QueryException $e) {
+              return response()->json(['error' => '資料庫錯誤：' . $e->getMessage()], 500);
+        } catch (Exception $e) {
+              return response()->json(['error' => '程式錯誤：' . $e->getMessage()], 500);
         }
+
         return view('promotion1.edit', compact('promotion1'))
                ->with(compact('results'))
                ->with(compact('products'))
@@ -63,22 +71,28 @@ class Promotion1Controller extends Controller
     public function update(Request $request, HpPromotion $promotion1)
     {
         $data = $request->all();
-        if (isset($data['paid'])) {
-            $data['remain'] = $promotion1->total-$data['paid'];
-        } else {
-            $data['remain'] = 0;
+        try {
+              if (isset($data['paid'])) {
+                  $data['remain'] = $promotion1->total-$data['paid'];
+              } else {
+                  $data['remain'] = 0;
+              }
+              if (($data['remain'] == 0) && ($promotion1->paid > 0)) {
+                 if (isset($data['product_id']) && $data['product_id'] != $promotion1->product_id) {
+                     $product = HpProduct::find($data['product_id']);
+                     $price = $product->price;
+                     //$total = $price * $promotion->amount;
+                     $data['remain'] = $promotion1->total-$promotion1->paid;
+                     $promotion1->update($data);
+                     return redirect()->route('promotion1.edit', compact('promotion1'));
+                 }
+              }
+              $promotion1->update($data);
+        } catch (QueryException $e) {
+              return response()->json(['error' => '資料庫錯誤：' . $e->getMessage()], 500);
+        } catch (Exception $e) {
+              return response()->json(['error' => '程式錯誤：' . $e->getMessage()], 500);
         }
-        if (($data['remain'] == 0) && ($promotion1->paid > 0)) {
-           if (isset($data['product_id']) && $data['product_id'] != $promotion1->product_id) {
-               $product = HpProduct::find($data['product_id']);
-               $price = $product->price;
-               //$total = $price * $promotion->amount;
-               $data['remain'] = $promotion1->total-$promotion1->paid;
-               $promotion1->update($data);
-               return redirect()->route('promotion1.edit', compact('promotion1'));
-           }
-        }
-        $promotion1->update($data);
 
         return redirect()->route('promotion1.edit', compact('promotion1'));
     }
@@ -90,6 +104,12 @@ class Promotion1Controller extends Controller
             return redirect()->route('eapplies.index');
         }
         foreach($ids as $id) {
+            $order_id = '91'.sprintf('%06d', $id);
+            $response = $this->gasImport($order_id);
+            $proms = json_decode($response, true);
+            if(!isset($proms['回傳結果'])) {
+                continue;
+            }
             $promotion = HpPromotion::find($id);
             $bundles = $this->createBundles($promotion);
             $data = $this->createFormArray($promotion, $bundles);
@@ -109,8 +129,14 @@ class Promotion1Controller extends Controller
     {
         $id = $request->input('id');
 
-        $promotion = HpPromotion::find($id);
+        $order_id = '91'.sprintf('%06d', $id);
+        $response = $this->gasImport($order_id);
+        $proms = json_decode($response, true);
+        if(!isset($proms['回傳結果'])) {
+            return redirect()->route('promotion1.index');
+        }
 
+        $promotion = HpPromotion::find($id);
         $bundles = $this->createBundles($promotion);
         $data = $this->createFormArray($promotion, $bundles);
         $response = $this->transfer(1, $data);
@@ -126,12 +152,24 @@ class Promotion1Controller extends Controller
 
     public function destroy(HpPromotion $promotion1)
     {
-        if ($promotion1->status == false) {
-            $promotion1->delete();
-        } else {
-            $promotion1->status = false;
-            $promotion1->save();
+        try {
+              if ($promotion1->status == false) {
+                  if ($promotion1->EcpayInfo != null) {
+                      $promotion1->EcpayInfo->delete();
+                  }
+                  if ($promotion1->EcpayResult == null) {
+                      $promotion1->delete();
+                  }
+              } else {
+                  $promotion1->status = false;
+                  $promotion1->save();
+              }
+        } catch (QueryException $e) {
+              return response()->json(['error' => '資料庫錯誤：' . $e->getMessage()], 500);
+        } catch (Exception $e) {
+              return response()->json(['error' => '程式錯誤：' . $e->getMessage()], 500);
         }
+
         return redirect()->route('promotion1.index');
     }
 
@@ -199,7 +237,7 @@ class Promotion1Controller extends Controller
                      '地址'          => $promotion->address,
                      '進件單位'      => $promotion->reseller->name,
                      '備註說明'      => $promotion->memo,
-                     '商品名稱'      => 'Z5000W AI智慧門鎖',
+                     '商品名稱'      => '商品一批',
                      '訂購數量'      => $promotion->amount,
                      '訂購方案'      => $promotion->product->paytype,
                      '收款方式'      => ($promotion->payment == 2) ? '綠界多元支付' : '其他',
@@ -257,6 +295,32 @@ class Promotion1Controller extends Controller
        curl_close($curl);
 
        return $response;
+    }
+
+    public function gasImport($orderid)
+    {
+        $curl = curl_init();
+
+        $url = config('gas.export_project_url').'?orderid='. $order_id;
+
+        curl_setopt_array($curl, array(
+               CURLOPT_URL => $url,
+               CURLOPT_RETURNTRANSFER => true,
+               CURLOPT_FOLLOWLOCATION => true,
+               CURLOPT_ENCODING => "",
+               CURLOPT_TIMEOUT => 30000,
+               CURLOPT_HTTP_VERSION => CURL_HTTP_VERSION_1_1,
+               CURLOPT_CUSTOMREQUEST => "GET",
+               CURLOPT_POSTFIELDS => null,
+               CURLOPT_HTTPHEADER => array(
+                   'Content-Type: application/json',
+              ),
+        ));
+
+        $response = curl_exec($curl);
+        $err = curl_error($curl);
+        curl_close($curl);
+        return $response;
     }
 
 }
